Implement password visibility toggle

+++ src/frontend/components/dashboard/ActiveChat.jsx
@@ -96,7 +96,7 @@ export default function ActiveChat({ socket, currentUser, activeChat, setActiveC
     if (msgDate >= startOfToday) return `Today, ${msgDate.toLocaleTimeString([], timeFormat)}`;
     else if (msgDate >= startOfYesterday) return `Yesterday, ${msgDate.toLocaleTimeString([], timeFormat)}`;
     else if (msgDate >= startOfWeekAgo) return `${msgDate.toLocaleDateString([], { weekday: 'long' })}, ${msgDate.toLocaleTimeString([], timeFormat)}`;
-    else return `${msgDate.toLocaleDateString([], { year: 'numeric', month: 'long', day: 'numeric' })}, ${msgDate.toLocaleTimeString([], timeFormat)}`;
+    else return `${msgDate.toLocaleDateString([], { year: 'numeric', month: 'short', day: 'numeric' })}, ${msgDate.toLocaleTimeString([], timeFormat)}`;
   }
 
   function capitalizeFirstLetter(str) {
